feat(attest): only format type strings when checked (#1065)
